+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/AppContext.kt
@@ -15,12 +15,9 @@ import com.facebook.react.uimanager.common.UIManagerType
 import expo.modules.adapters.react.NativeModulesProxy
 import expo.modules.core.errors.ContextDestroyedException
 import expo.modules.core.interfaces.ActivityProvider
-import expo.modules.interfaces.camera.CameraViewInterface
 import expo.modules.interfaces.constants.ConstantsInterface
-import expo.modules.interfaces.font.FontManagerInterface
 import expo.modules.interfaces.imageloader.ImageLoaderInterface
 import expo.modules.interfaces.permissions.Permissions
-import expo.modules.interfaces.taskManager.TaskManagerInterface
 import expo.modules.kotlin.activityresult.ActivityResultsManager
 import expo.modules.kotlin.activityresult.DefaultAppContextActivityResultCaller
 import expo.modules.kotlin.defaultmodules.ErrorManagerModule
@@ -187,24 +184,6 @@ class AppContext(
    * Provides access to the image loader from the legacy module registry
    */
   val imageLoader: ImageLoaderInterface?
-    get() = legacyModule()
-
-  /**
-   * Provides access to the camera view manager from the legacy module registry
-   */
-  val camera: CameraViewInterface?
-    get() = legacyModule()
-
-  /**
-   * Provides access to the font manager from the legacy module registry
-   */
-  val font: FontManagerInterface?
-    get() = legacyModule()
-
-  /**
-   * Provides access to the task manager from the legacy module registry
-   */
-  val taskManager: TaskManagerInterface?
     get() = legacyModule()
 
   /**